requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
Labelme2YOLOv8==0.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.5.0
